split items into items and helpers
